requirements.txt
websocket-client==1.1.0
